WIP [ci skip]

+++ databox/client/src/components/Media/CollectionMenuItem.tsx
@@ -2,13 +2,17 @@ import {PureComponent, MouseEvent} from "react";
 import {Collection} from "../../types";
 import {getCollections} from "../../api/collection";
 import {ReactComponent as ArrowDownImg} from '../../images/icons/arrow-down.svg';
+import {ReactComponent as EditImg} from '../../images/icons/edit.svg';
+import {ReactComponent as TrashImg} from '../../images/icons/trash.svg';
 import Icon from "../ui/Icon";
 import Button from "../ui/Button";
 import apiClient from "../../api/api-client";
+import EditCollection from "./Collection/EditCollection";
 
 type State = {
     collections?: Collection[],
     expanded: boolean,
+    editing: boolean,
 }
 
 export type CollectionMenuItemProps = {
@@ -21,6 +25,7 @@ export type CollectionMenuItemProps = {
 export default class CollectionMenuItem extends PureComponent<CollectionMenuItemProps, State> {
     state: State = {
         expanded: false,
+        editing: false,
     };
 
     expandCollection = async (force = false): Promise<void> => {
@@ -52,8 +57,20 @@ export default class CollectionMenuItem extends PureComponent<CollectionMenuItem
         this.expandCollection();
     }
 
-    delete = () => {
-        apiClient.delete(`/collections/${this.props.id}`);
+    edit = (e: MouseEvent): void => {
+        e.stopPropagation();
+        this.setState({editing: true});
+    }
+
+    closeEdit = () => {
+        this.setState({editing: false});
+    }
+
+    delete = (e: MouseEvent): void => {
+        e.stopPropagation();
+        if (window.confirm(`Delete? Really?`)) {
+            apiClient.delete(`/collections/${this.props.id}`);
+        }
     }
 
     render() {
@@ -65,6 +82,7 @@ export default class CollectionMenuItem extends PureComponent<CollectionMenuItem
             capabilities,
             level,
         } = this.props;
+        const {editing} = this.state;
 
         const selected = selectedPath === absolutePath;
         const currentInSelectedHierarchy = selectedPath && selectedPath.startsWith(absolutePath);
@@ -87,21 +105,31 @@ export default class CollectionMenuItem extends PureComponent<CollectionMenuItem
                 </div>
                 <div className="actions">
                     {capabilities.canEdit ? <Button
-                        className={'btn-secondary'}
-                        disabled={true}
-                    >E</Button> : ''}
+                        size={"sm"}
+                        onClick={this.edit}
+                    ><Icon
+                        component={EditImg}/></Button> : ''}
                     {capabilities.canDelete ? <Button
+                        size={"sm"}
                         onClick={this.delete}
-                        className={'btn-danger'}
-                    >D</Button> : ''}
+                    ><Icon
+                        component={TrashImg}
+                    /></Button> : ''}
                 </div>
                 {children!.length > 0 ? <div
                     className="expand"
                     onClick={this.onExpandClick}
                 >
-                    <Icon component={ArrowDownImg}/>
+                    <Icon
+                        variant={'xs'}
+                        component={ArrowDownImg}
+                    />
                 </div> : ''}
             </div>
+            {editing ? <EditCollection
+                id={this.props.id}
+                onClose={this.closeEdit}
+            /> : ''}
             {this.renderChildren()}
         </div>
     }
@@ -115,6 +143,7 @@ export default class CollectionMenuItem extends PureComponent<CollectionMenuItem
         return <div className="sub-colls">
             {collections.map(c => <CollectionMenuItem
                 {...c}
+                key={c.id}
                 absolutePath={`${this.props.absolutePath}/${c.id}`}
                 selectedPath={this.props.selectedPath}
                 onClick={this.props.onClick}
